Muevo módulos de indexing fuera de la app de api
    refs:#151

    refs:#151
