--- series_tiempo_ar_api/apps/api/common/__init__.py
@@ -1,1 +1,0 @@
-#! coding: utf-8
