+++ series_tiempo_ar_api/apps/api/tests/reader_tests.py
@@ -6,15 +6,15 @@ from django.conf import settings
 from django.test import TestCase
 from elasticsearch_dsl import Search
 from series_tiempo_ar.search import get_time_series_distributions
-
-from series_tiempo_ar_api.apps.api.indexing.database_loader import \
+from series_tiempo_ar_api.libs.indexing.catalog_reader import index_catalog
+from series_tiempo_ar_api.libs.indexing.database_loader import \
     DatabaseLoader
-from series_tiempo_ar_api.apps.api.indexing.distribution_indexer import DistributionIndexer
-from series_tiempo_ar_api.apps.api.indexing.scraping import Scraper
+from series_tiempo_ar_api.libs.indexing.distribution_indexer import DistributionIndexer
+
 from series_tiempo_ar_api.apps.api.models import Distribution, Field, Catalog
-from series_tiempo_ar_api.apps.api.indexing.catalog_reader import index_catalog
 from series_tiempo_ar_api.apps.api.tests import setup_database
 from series_tiempo_ar_api.libs.indexing.elastic import ElasticInstance
+from series_tiempo_ar_api.libs.indexing.scraping import Scraper
 
 SAMPLES_DIR = os.path.join(os.path.dirname(__file__), 'samples')
 CATALOG_ID = 'test_catalog'
