+++ series_tiempo_ar_api/apps/api/tests/support/generate_data.py
@@ -9,11 +9,11 @@ from dateutil.relativedelta import relativedelta
 from django.conf import settings
 from elasticsearch.helpers import parallel_bulk
 
-from series_tiempo_ar_api.apps.api.common import operations
+from series_tiempo_ar_api.libs.indexing import operations
 from series_tiempo_ar_api.apps.api.helpers import interval_to_freq_pandas
-from series_tiempo_ar_api.apps.api.indexing.constants import INDEX_CREATION_BODY, \
-    FORCE_MERGE_SEGMENTS
 from series_tiempo_ar_api.apps.api.query.constants import COLLAPSE_INTERVALS
+from series_tiempo_ar_api.libs.indexing.constants import INDEX_CREATION_BODY, \
+    FORCE_MERGE_SEGMENTS
 from series_tiempo_ar_api.libs.indexing.elastic import ElasticInstance
 
 DATA_FILE_NAME = 'data.csv'
