+++ series_tiempo_ar_api/apps/management/views.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.shortcuts import render
-
 # Create your views here.
