+++ series_tiempo_ar_api/libs/indexing/constants.py
@@ -1,6 +1,25 @@
 #! coding: utf-8
 from django.conf import settings
-from ..common.constants import VALUE, CHANGE, PCT_CHANGE, CHANGE_YEAR_AGO, PCT_CHANGE_YEAR_AGO
+from pandas.tseries.offsets import YearBegin, QuarterBegin, MonthBegin, Day
+
+
+# Transformaciones
+VALUE = 'value'
+CHANGE = 'change'
+PCT_CHANGE = 'percent_change'
+CHANGE_YEAR_AGO = 'change_a_year_ago'
+PCT_CHANGE_YEAR_AGO = 'percent_change_a_year_ago'
+
+# Pandas freqs
+PANDAS_YEAR = YearBegin()
+PANDAS_SEMESTER = MonthBegin(6)
+PANDAS_QUARTER = QuarterBegin(startingMonth=1)
+PANDAS_MONTH = MonthBegin()
+PANDAS_WEEK = Day(7)
+PANDAS_DAY = Day()
+
+# Frecuencias *en orden* de mayor a menor
+PANDAS_FREQS = [PANDAS_YEAR, PANDAS_SEMESTER, PANDAS_QUARTER, PANDAS_MONTH, PANDAS_WEEK, PANDAS_DAY]
 
 IDENTIFIER = "identifier"
 DATASET_IDENTIFIER = "dataset_identifier"
